Fix integer division truncating the ripple size during the panic swipe
rippleSize and yMaxTranslation are both ints, so k was computed with
integer division and truncated to 0 whenever yMaxTranslation was odd.
The line right above it already casts to float for the same kind of
division (float v = (float) yCurrentTranslation / yMaxTranslation),
this one line was just missed. Cast rippleSize to float before
dividing so k keeps its fractional value.

+++ app/src/main/java/info/guardianproject/ripple/PanicActivity.java
@@ -151,7 +151,7 @@ public class PanicActivity extends Activity implements OnTouchListener {
 
                     int rippleSize = yMaxTranslation / 2;
                     if (yCurrentTranslation > rippleSize) {
-                        float k = rippleSize / (yMaxTranslation - rippleSize);
+                        float k = (float) rippleSize / (yMaxTranslation - rippleSize);
                         mRipples.setSize((yCurrentTranslation - rippleSize) * k);
                         mRipples.invalidate();
                     }
